Move UnzipTransform to a top-level type
