+++ subprojects/language-native/src/main/java/org/gradle/language/plugins/NativeBasePlugin.java
@@ -16,14 +16,11 @@
 
 package org.gradle.language.plugins;
 
-import com.google.common.io.Files;
-import org.apache.commons.io.IOUtils;
 import org.gradle.api.*;
 import org.gradle.api.artifacts.Configuration;
 import org.gradle.api.artifacts.ConfigurationContainer;
 import org.gradle.api.artifacts.ModuleVersionIdentifier;
 import org.gradle.api.artifacts.dsl.DependencyHandler;
-import org.gradle.api.artifacts.transform.ArtifactTransform;
 import org.gradle.api.attributes.Attribute;
 import org.gradle.api.attributes.AttributeCompatibilityRule;
 import org.gradle.api.attributes.AttributeContainer;
@@ -39,6 +36,7 @@ import org.gradle.api.file.DirectoryProperty;
 import org.gradle.api.file.FileCollection;
 import org.gradle.api.file.RegularFile;
 import org.gradle.api.internal.artifacts.ArtifactAttributes;
+import org.gradle.api.internal.artifacts.transform.UnzipTransform;
 import org.gradle.api.model.ObjectFactory;
 import org.gradle.api.plugins.ExtensionContainer;
 import org.gradle.api.provider.Provider;
@@ -49,7 +47,6 @@ import org.gradle.api.publish.maven.internal.publisher.MutableMavenProjectIdenti
 import org.gradle.api.tasks.TaskContainer;
 import org.gradle.api.tasks.TaskProvider;
 import org.gradle.internal.Cast;
-import org.gradle.internal.UncheckedException;
 import org.gradle.language.ComponentWithBinaries;
 import org.gradle.language.ComponentWithOutputs;
 import org.gradle.language.ComponentWithTargetMachines;
@@ -79,19 +76,10 @@ import org.gradle.nativeplatform.toolchain.NativeToolChain;
 import org.gradle.nativeplatform.toolchain.internal.PlatformToolProvider;
 
 import javax.inject.Inject;
-import java.io.BufferedInputStream;
-import java.io.File;
-import java.io.FileInputStream;
-import java.io.FileOutputStream;
-import java.io.IOException;
 import java.util.Collections;
-import java.util.List;
 import java.util.Set;
 import java.util.concurrent.Callable;
-import java.util.zip.ZipEntry;
-import java.util.zip.ZipInputStream;
 
-import static org.apache.commons.io.FilenameUtils.removeExtension;
 import static org.gradle.api.artifacts.type.ArtifactTypeDefinition.C_PLUS_PLUS_API_DIRECTORY;
 import static org.gradle.api.artifacts.type.ArtifactTypeDefinition.C_PLUS_PLUS_API_ZIP_TYPE;
 import static org.gradle.api.artifacts.type.ArtifactTypeDefinition.DIRECTORY;
@@ -477,45 +465,6 @@ public class NativeBasePlugin implements Plugin<Project> {
             if (C_PLUS_PLUS_API_DIRECTORY.equals(details.getConsumerValue())
                     && DIRECTORY.equals(details.getProducerValue())) {
                 details.compatible();
-            }
-        }
-    }
-
-    private static class UnzipTransform extends ArtifactTransform {
-        @Inject
-        UnzipTransform() { }
-
-        @Override
-        public List<File> transform(File zippedFile) {
-            String unzippedDirName = removeExtension(zippedFile.getName());
-            File unzipDir = new File(getOutputDirectory(), unzippedDirName);
-            try {
-                unzipTo(zippedFile, unzipDir);
-            } catch (IOException e) {
-                throw UncheckedException.throwAsUncheckedException(e);
-            }
-            return Collections.singletonList(unzipDir);
-        }
-
-        private void unzipTo(File headersZip, File unzipDir) throws IOException {
-            ZipInputStream inputStream = new ZipInputStream(new BufferedInputStream(new FileInputStream(headersZip)));
-            try {
-                ZipEntry entry;
-                while ((entry = inputStream.getNextEntry()) != null) {
-                    if (entry.isDirectory()) {
-                        continue;
-                    }
-                    File outFile = new File(unzipDir, entry.getName());
-                    Files.createParentDirs(outFile);
-                    FileOutputStream outputStream = new FileOutputStream(outFile);
-                    try {
-                        IOUtils.copyLarge(inputStream, outputStream);
-                    } finally {
-                        outputStream.close();
-                    }
-                }
-            } finally {
-                inputStream.close();
             }
         }
     }
